// model/config.ts
export interface MovieResponse{
    mvid: number;
    title: string;
    poster: string;
    year: string;
    rating: string;
    runtime: string;
    plot_movie: string;
}

export interface PersonResponse{
    pid: number;
    name: string;
    img: string;
    age: string;
    detail: string;
}

export interface SeacrhResponse{
    mvid: number;
    title: string;
    poster: string;
    year: string;
    rating: string;
    runtime: string;
    plot_movie: string;
    pid: number;
    name: string;
    img: string;
    age: string;
    detail: string;
}